Resolve "创建索引可以指定不同的ttltype"

Closes #374

See merge request FeatureEngineering/rtidb!492

+++ src/sdk/mini_cluster_bm.cc
@@ -280,7 +280,7 @@ void BM_BatchRequestQuery(benchmark::State& state, fesql::sqlcase::SQLCase& sql_
                 continue;
             }
             // don't repeat request in debug mode
-            for (size_t repeat_idx = 1; repeat_idx < sql_case.batch_request_.repeat_; repeat_idx++) {
+            for (int repeat_idx = 1; repeat_idx < sql_case.batch_request_.repeat_; repeat_idx++) {
                 ASSERT_TRUE(row_batch->AddRow(request_row));
             }
         }
